Create checkAndUncheck function in a separate file and exported to index.js for code modularization
