+++ src/index.js
@@ -1,5 +1,6 @@
 import './main.css';
 import clearListContainer from './utils/clear-container.js';
+import checkAndUncheckTodo from './utils/check-uncheck-todo.js';
 
 const listContainer = document.querySelector('[data-lists]');
 const newTodoForm = document.querySelector('[data-new-todo-form]');
@@ -53,30 +54,7 @@ const handleRender = () => {
     }
     listContainer.appendChild(listElement);
   });
-
-  document.querySelectorAll('.check-box').forEach((check) => {
-    check.addEventListener('change', (e) => {
-      const todos = JSON.parse(localStorage.getItem(localStorageTodos));
-      const todoId = e.target.closest('li').dataset.todoItem;
-      const foundTodo = todos.find((todo) => todo.id === todoId);
-
-      if (foundTodo.completed) {
-        e.target.parentNode.nextSibling.nextElementSibling
-          .querySelector('.todo-desc')
-          .classList.remove('linethrough');
-        foundTodo.completed = false;
-      } else {
-        e.target.parentNode.nextSibling.nextElementSibling
-          .querySelector('.todo-desc')
-          .classList.add('linethrough');
-        foundTodo.completed = true;
-      }
-      const filteredTodo = todos.filter((todo) => todo.id !== todoId);
-      filteredTodo.splice(foundTodo.index - 1, 0, foundTodo);
-      localStorage.setItem(localStorageTodos, JSON.stringify(filteredTodo));
-    });
-  });
-  
+  checkAndUncheckTodo(todos, localStorageTodos);
 };
 
 const handleSaveAndRender = () => {
